FEAT : mim.f90 is modified
Definition of Q_exist is changed and Qcomps_exist is newly defined.
Output parameters related to the total diabatic heating are estimated
by diabaticHeating() in this version.
If INPUT_Q_FILENAME=='' and ttswr, ttlwr, lrghr, cnvhr, and vdfhr are
inputted, Q-related outputs are computed by the sum of output parameters
for these five inputs.