FEAT : New module az_latprof.f90 is added
This module computes the latitudinal profile of zonal available
potential energy.
The original one contains the term that contributes to the global mean
available potential energy, and the term is dominant.
In this module such term is excluded.
The subroutine has not already made, this commit is just to save the file.